Double

+++ ASDF/JavaApplication27/src/javaapplication27/JavaApplication27.java
@@ -19,7 +19,7 @@ public class JavaApplication27 {
        
         JOptionPane.showMessageDialog(null, "Para calcular el área del pentágono:");
         
-       double a = Integer.parseInt(JOptionPane.showInputDialog("Introduzca la apotema"));
+       double a = Double.parseDouble(JOptionPane.showInputDialog("Introduzca la apotema"));
        double l = Integer.parseInt(JOptionPane.showInputDialog("Introduce lado"));
        
        JOptionPane.showMessageDialog(null, "El area del pentágono de lado " + l + "cm y apotema " + a + "cm es igual a " + l*5*a/2 + "cm².");
